requirements.txt
# use python 3.8.20

ydata-profiling>=4.1.2
seaborn>=0.11.2
matplotlib>=3.3.4
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.1
 
scikit-learn>=1.0.2
pandas>=1.3.0
numpy>=1.20.0

imbalanced-learn>=0.10.1
 
xgboost>=1.5.0
lightgbm>=3.3.2
 
evidently>=0.4.14
 
mlflow>=2.4.1

shap

python-dotenv
